requirements.txt
django
djangorestframework
drf-spectacular
psycopg2-binary
django-environ
